requirements.txt
pyTelegramBotAPI==4.14.0
gspread==5.12.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
pytz==2023.3
openai==1.98.0
pandas==2.1.4
tinydb==4.8.0
apscheduler==3.10.4
qrcode==7.4.2
Pillow==11.3.0
python-dotenv==1.0.0
requests==2.32.4
typing_extensions==4.14.1
tzlocal==5.3.1

# PostgreSQL
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
